refactor(tentacle): extract PR automation seam
