+++ hooks/pre-commit
@@ -125,6 +125,7 @@ def in_python_cleanliness_surface(path: str) -> bool:
         "tentacle.py",
         "_tentacle_core.py",
         "_tentacle_goal.py",
+        "_tentacle_pr.py",
         "checkpoint-diff.py",
         "checkpoint-restore.py",
         "checkpoint-save.py",
