+++ tests/test_quality_gates.py
@@ -848,6 +848,7 @@ CI_RUFF_FILES = [
     "tentacle.py",
     "_tentacle_core.py",
     "_tentacle_goal.py",
+    "_tentacle_pr.py",
     "checkpoint-diff.py",
     "checkpoint-restore.py",
     "checkpoint-save.py",
@@ -1368,6 +1369,7 @@ def test_architecture_md_ruff_surface():
         "tentacle.py",
         "_tentacle_core.py",
         "_tentacle_goal.py",
+        "_tentacle_pr.py",
         "tests/test_browse_search_v2.py",
         "browse/",
         "hooks/",
